Started files for the input side of the GUI
Currently, there is a separate html file for each input page — home,
personal, academic, and course load.  we can combine these into an SPA
but it will be easier to divide/conquer initially if they are separate.

created a stylesheet and js file for the whole project, we should keep
these as separate from the html files as possible seeing as we may be
combining files later on.